requirements.txt
requests==2.31.0
certifi
duckdb==0.10.1
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scipy==1.12.0
statsmodels==0.14.1
streamlit==1.32.2
plotly==5.20.0
tqdm==4.66.2
loguru==0.7.2
python-dotenv==1.0.1
pyarrow==15.0.2
scikit-learn==1.4.1.post1
